refactor: Recoil 추가 분리 및 회원정보 조회 로직 분리
