+++ hooks/useOnPlay.ts
@@ -4,7 +4,7 @@ import usePlayer from "./usePlayer";
 //import useSubscribeModal from "./useSubscribeModal";
 import useAuthModal from "./useAuthModal";
 import {useRecoilState} from "recoil";
-import {loginState} from "@/store/store";
+import {loginState} from "@/providers/RecoilContextProvider";
 
 //basically when use clicks the button play current song based on id but also create a playlist of the played songs
 const useOnPlay = (songs: Song[]) => {
