+++ providers/RecoilContextProvider.tsx
@@ -1,4 +1,7 @@
-import {atom, RecoilState} from "recoil";
+'use client';
+
+import {RecoilRoot, atom, RecoilState} from 'recoil';
+import React from 'react';
 import {UserInfos} from "@/types";
 
 export const baseUrl = atom({
@@ -20,4 +23,6 @@ export const userInfoState : RecoilState<UserInfos> = atom({
     profileImageUrl: "",
   }
 });
-
+export default function RecoilContextProvider({ children }: { children: React.ReactNode }) {
+  return <RecoilRoot>{children}</RecoilRoot>;
+}
